fix for sigmoid lowering
