+++ test/AotCompile/model_loader_lib.py
@@ -98,3 +98,27 @@ def broadcast_add_mixed_ranks_loader() -> TorchLoaderOutput:
         inputs=[x, y],
         constraints=constraints,
     )
+
+
+def sigmoid_loader() -> TorchLoaderOutput:
+    class Sigmoid(torch.nn.Module):
+        def __init__(self):
+            super().__init__()
+
+        def forward(self, x: torch.Tensor) -> torch.Tensor:
+            return torch.sigmoid(x)
+
+    # Sample inputs
+    x = torch.randn(2, 3)
+
+    # Dynamic dim constraints
+    constraints = [
+        dynamic_dim(x, 0),
+        dynamic_dim(x, 1),
+    ]
+
+    return TorchLoaderOutput(
+        model=Sigmoid(),
+        inputs=[x],
+        constraints=constraints,
+    )
